Unset CARGO_ENCODED_RUSTFLAGS from binding-generator cargo command.
The ENCODED_RUSTFLAGS are both read and written by cargo. This
potentially generates conflicts when cross compiling and some flags
are set for target.<triple>.rustflags in the current target
configuration. These should never be used for the bindings-generator,
as it will always be built for the host target. If some extra flags
are needed, they can be set in RUSTFLAGS.